move route profile to users
